backend/requirements/base.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
psycopg2-binary==2.9.9
reportlab==4.2.2
gunicorn==23.0.0
whitenoise==6.7.0
